Checking in for release

+++ lib/tasks/usman/data.rake
@@ -9,7 +9,7 @@ namespace 'usman' do
       desc "Import all data in sequence"
       task 'all' => :environment do
 
-        import_list = ["users", "features", "permissions", "roles"]
+        import_list = ["users", "permissions"]
         
         import_list.each do |item|
           print "Importing #{item.titleize} \t".yellow
@@ -25,17 +25,20 @@ namespace 'usman' do
         puts " "
       end
 
-      ["Feature", "User", "Role", "Permission"].each do |cls_name|
+      ["User", "Permission"].each do |cls_name|
         name = cls_name.underscore.pluralize
         desc "Import #{cls_name.pluralize}"
         task name => :environment do
           verbose = true
           verbose = false if ["false", "f","0","no","n"].include?(ENV["verbose"].to_s.downcase.strip)
+
+          destroy_all = false
+          destroy_all = true if ["true", "t","1","yes","y"].include?(ENV["destroy_all"].to_s.downcase.strip)
           
           path = Rails.root.join('db', 'data', "#{cls_name.constantize.table_name}.csv")
           path = Usman::Engine.root.join('db', 'data', "#{cls_name.constantize.table_name}.csv") unless File.exists?(path)
           
-          cls_name.constantize.destroy_all
+          cls_name.constantize.destroy_all if destroy_all
           cls_name.constantize.import_data_file(path, true, verbose)
           puts "Importing Completed".green if verbose
         end
@@ -46,7 +49,7 @@ namespace 'usman' do
         desc "Import all dummy data in sequence"
         task 'all' => :environment do
 
-          import_list = ["users", "features", "permissions", "roles"]
+          import_list = ["users", "permissions"]
           
           import_list.each do |item|
             print "Loading #{item.split(':').last.titleize} \t".yellow
@@ -62,17 +65,20 @@ namespace 'usman' do
           puts " "
         end
 
-        ["Feature", "User", "Role", "Permission"].each do |cls_name|
+        ["User", "Permission"].each do |cls_name|
           name = cls_name.underscore.pluralize
           desc "Load Dummy #{cls_name.pluralize}"
           task name => :environment do
             verbose = true
             verbose = false if ["false", "f","0","no","n"].include?(ENV["verbose"].to_s.downcase.strip)
+
+            destroy_all = false
+            destroy_all = true if ["true", "t","1","yes","y"].include?(ENV["destroy_all"].to_s.downcase.strip)
             
             path = Rails.root.join('db', 'data', 'dummy', "#{cls_name.constantize.table_name}.csv")
             path = Usman::Engine.root.join('db', 'data', 'dummy', "#{cls_name.constantize.table_name}.csv") unless File.exists?(path)
 
-            cls_name.constantize.destroy_all
+            cls_name.constantize.destroy_all if destroy_all
             cls_name.constantize.import_data_file(path, true, verbose)
             puts "Importing Completed".green if verbose
           end
